[TRY-48] RENAME: Subdivide security package
